requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
firebase-admin==3.0.0
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.2
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
